[CMLG-001] Enabled responsive scroll of the table grid
Issue:
When scrolling horizontally the entrie webpage moves, when we only want the table to in a smaller screen.

Solution:
Changed the css for the container for the table tag to overflow-x: auto !important.
Fixed the problem with z-index of the table headers.

Risk:
The !important flag may cause issues, but entirely sure.

Reviewed by: Yujia, Kevin