[CMLG-007] fix the first column (#8)
Issue:

The background colour of each row doesn't go all the way to the fixed columns.
There is a gap between first and second column.
title row looks exactly the same as the body of the table.

Solution:

Fix only the first column and leave the second column unfixed.
Change the background colour of the first column, so the background colour across each row is consistent.
Give the title a different background colour and bold the title text.

Notes:
To fix the position of the second column, we need to know the width of first column during runtime. This is difficult to achieve now but it is something that user wants. Therefore, we will still need to implement it in the future.

Risk: /

Reviewed by:
Annie, Eileen, Dave, Yujia, Kevin