[trial] Trying to see if it is possible to create a table using pure html
Issue:

Calling the API locally can cause the CROS issue.

Solution:

Downloading an extension on browser can solve this problem when developing locally, trying to see if the website still has this problem after deploying to the server.

Risk:
none, this branch will not be used for real development

Reviewed by:
Emily