[CMLG-011] clean the file structure (#11)
Issue:

All files are written under the src folder, it is hard to find a specific file quickly.

Solution:

Creating a component folder to hold all react components created. The search page is created to render Table, SearchBar and SelectColumn components (yet to be implemented).
A bootstrap CSS link is also added to allow the use of the bootstrap library in frontend development.

Risk:
No risk.
Some files, e.g. TranslationTable.js, are not deleted at this point in case developers need to refer to them. These should be deleted in the future.

Reviewed by:
Yujia, Kevin, Annie, Eileen, Dave