[CMLG-011] Sort table when clicking on any header
Issue:

When the user clicks on a header, the table should be sorted in increasing order based on the column that user clicks. When the user clicks the same header again, the table should be sorted in decreasing order.

Solution:

Creating a function to sort the data when user clicks a header. An arrow will appear beside the header name to show the current sorting order.
Because each column needs to remember its current sorting order, another state is created in the table to store the sorting order of each column.

Risk:

Comparison of words is done using the function Intl.Collator(). Part of the sorting results are compared with results using excel sheet to test the correctness of the program. Among all languages, "Arebic" doesn't produce the same result. However, the result is the same as the one in MVP.

Words start with '(' will be considered as smaller than words without brackets. For example, although 'y' comes after 'g', word (yìdàlì) is smaller than 'guójì' because '(' has a smaller ASCII number. We can probably solve this problem by standardise the database.

Reviewed by:
Annie, Yujia, Kevin, Dave, Eileen