[CMLG - 013] Select column change the passed column to search page including all the language with their state.
Issue: We only pass the selected language to the search page component previously, but we need the state of all the languages.

Solution: Initialise the state of each language as a 2-dimensional array, when we click a checkbox, update the state of the selected language, and pass the array to the search page.

Risk: Unknown.

Reviewed by: Yujia, Emily, Dave, Kevin, Annie