[CMLG-013] Select Columns
Issues: We need a component to display all the languages and handle the change of user selection (i.e. select/unselect the language). 

Solution: Initialize the state in SearchPage component and pass the state to SelectCol component to handle the change, then pass back to SearchPage.

Risk: Unknown. Need to test by table component.

Reviewed by: Yujia, Emily, Annie, Kavin, Dave