[CMLG-011] change Table component to not store the name of each header
Issue:

SearchPage is passing names of headers to the Table compoent, Table component no longer needs to store this information.

The header name "Portugues" and "Turkce" are not the same as the one in the excel sheet that the user provides.

Solution:

Table component is now displaying the headers of the table using the props passed by the SearchPage component.

Changing Portugues to Português and Turkce to Türkçe in order to match the excel sheet.

Risk:
/

Reviewed by:
Annie, Eileen, Yujia, Kevin, Dave