[CMLG-011] only show columns that the user selects
Issue:

When the user select/unselect a column, the table be modified to display only columns that the user selects.

Solution:

Column selected status is passed from the SearchPage component to the Table component. If a column is selected, the "display" css of all table cells in that column is changed to "table-cell", otherwise it is set to be "none".

Risk:
/

Reviewed by:
Annie, Eileen, Yujia, Kevin, Dave