[CMLG-012] Styling the SearchBar and SelectCol (#21)
* Change the icon position to the right

* [CMLG-011] Sort table when user clicks the header  (#17)

Issue:

When the user clicks on a header, the table should be sorted in increasing order based on the column that the user clicks. When the user clicks the same header again, the table should be sorted in decreasing order.

Solution:

Creating a function to sort the data when the user clicks a header. An arrow will appear beside the header name to show the current sorting order.
Because each column needs to remember its current sorting order, another state is created in the table to store the sorting order of each column.

Risk:

Comparison of words is done using the function Intl.Collator(). Part of the sorting results is compared with results using an excel sheet to test the correctness of the program. Among all languages, "Arabic" doesn't produce the same result. However, the result is the same as the one in MVP.

Words starting with '(' will be considered as smaller than words without brackets. For example, although 'y' comes after 'g', word (yìdàlì) is smaller than 'guójì' because '(' has a smaller ASCII number. We can probably solve this problem by standardising the database.

Reviewed by:
Annie, Yujia, Kevin, Dave, Eileen

* Horizontally align the search bar and the select column button.

* [CMLG-011] Sort table when clicking on any header

Issue:

When the user clicks on a header, the table should be sorted in increasing order based on the column that user clicks. When the user clicks the same header again, the table should be sorted in decreasing order.

Solution:

Creating a function to sort the data when user clicks a header. An arrow will appear beside the header name to show the current sorting order.
Because each column needs to remember its current sorting order, another state is created in the table to store the sorting order of each column.

Risk:

Comparison of words is done using the function Intl.Collator(). Part of the sorting results are compared with results using excel sheet to test the correctness of the program. Among all languages, "Arebic" doesn't produce the same result. However, the result is the same as the one in MVP.

Words start with '(' will be considered as smaller than words without brackets. For example, although 'y' comes after 'g', word (yìdàlì) is smaller than 'guójì' because '(' has a smaller ASCII number. We can probably solve this problem by standardise the database.

Reviewed by:
Annie, Yujia, Kevin, Dave, Eileen

* [CMLG-011] only show columns that the user selects

Issue:

When the user select/unselect a column, the table be modified to display only columns that the user selects.

Solution:

Column selected status is passed from the SearchPage component to the Table component. If a column is selected, the "display" css of all table cells in that column is changed to "table-cell", otherwise it is set to be "none".

Risk: /

Reviewed by:
Annie, Eileen, Yujia, Kevin, Dave