[CMLG-011] Instant Search (#22)
Issue:
We wanted to implement instant search, which works like this: the words that the user is searching for (input into the search bar) are displayed on the page real time. Clicking the search icon after typing the search word (e.g. Google) is not required as search operation happens real time.
Also, we changed our backend API so our app didn't work properly.

Solution:
Used componentDidUpdate() to change the URL that is fetched every time when the search word is changed. For example, when searching for the word virus, the URL will change from https://cmlgbackend.wdcc.co.nz/api/translations?sequence=1596773281500&word= to https://cmlgbackend.wdcc.co.nz/api/translations?sequence=1596773281500&word=virus.
In order to make the app work with the new backend API, we changed the URL that is fetched, and made sure that the data that gets processed to be displayed on the website does not include the sequence numbers because we don't want to display sequence numbers.

Risk:
No risk.

Reviewed by:
Emily, Yujia, Annie, Eileen, Kevin.