[CMLG-011] fix the displaying data issue with new API
Issue:

The table doesn't have a word property so the search word is not passed to Table.
When the website first loads, word is null but we don't want to pass null as the search word to the API. Instead, we want the API to returns the first two rows.

Solution:

Pass a property words to the table component.
Check if the website is loading, if it is, call /translations?sequence=xxx end point instead of /translations?sequence=xxx?word=null

Risk:
/

Reviewed by:
Annie, Dave, Eileen, Yujia, Kevin