[CMLG-012] Responsive search bar (#28)
Issue:
The search bar is not responsive. Some part of the border will be covered by the background of placeholder if the screen size is small.

Solution:
Change the placeholder text when the screen size is smaller than 900px. Also, change the size of the search bar for the mobile phone screen.

Risk:
Unknown.

Reviewed by:
Annie, Eileen, Dave, Emily, Yujia