[CMLG-011] put the scroll bar on table only
Issue:

Scrolling bar is put on the browser, when the user scrolls the page to the right, all components including the search bar and select component are scrolled.

Solution:

Display the searchPage in "grid", the searchPage component contains two rows. Searchbar and select components are in the first row. Table is in the second row. Change the overflow of table to be "scroll".

Risk:

When scrolling the table vertically, the header moves as well.

Reviewed by:
Annie, Dave, Eileen, Yujia, Kevin