[CMLG-011] only display most recent search result (#27)
Issue:

Frontend gets responses in random order, the frontend needs to decide which response to process and put on the table. For example, when the user searches for "u" and "p", the response for "up" comes first. When the response for "u" arrives, the frontend needs to realise this response is no longer needed and delete it.

Solution:

Because the response contains a sequence number, the Table component contains the sequence of the current displayed response. When receiving the next response, the Table component looks at the sequence number, and only displays the data if its sequence is larger than the current one.

Risk:
/

Reviewed by:
Annie, Eileen, Dave, Kevin, Yujia