[CMLG-011] fix responsive issue when sort columns
Issue:

When sorting columns, there is a small arrow showing beside the header clicked, indicating the sorting order. If the width of the header is not enough, the arrow goes to the next line.

Solution:

Add CSS style "white-space: nowrap" to force the arrow to be on the same line.

Risk:

As the width of the header is not enough for the arrow to be on the same line, when the arrow appears, the width of the header will get automatically larger to fit both words and arrow.

Reviewed by:
Annie, Eileen, Kevin, Dave, Yujia