[CMLG-011] Put the scroll bar on table only (#29)
Issue:

Scrolling bar is put on the browser, when the user scrolls the page to the right, all components including the search bar and select component are scrolled.

Solution:

The table cannot have "overflow" CSS on it. Therefore, change the display of searchPage to be "grid", the search page component contains two rows. Searchbar and select components are in the first row. The table is in the second row. Change the overflow of the div around the table to be "scroll".

Risk:

When scrolling the table vertically, the header moves as well.

Reviewed by:

Annie, Dave, Eileen, Yujia, Kevin