[CMLG-011] get width of first table column dynamically
Issue:

We need to fix Chinese and English columns when scorlling the table horizontally.

Solution:

Create a reference on "Chinese" header. this.chineseHeader stores this Chinese header element. To get the width of it, call this.chineseHeader.current.offsetWidth.
Using this value as the left CSS on each elements in second column can fixed its position, however, if there is any changes in the offsetWidth, the second columns will not be updated automatically.

Risk:
Functionality of fixed columns is not achieved at this point.

Reviewed by:
Annie, Yujia, Kevin, Eileen, Dave