[CMLG-011] Fix left two columns
Issue:
User wants Chinese and English columns fixed on the left.

Solution: 
Use sticky to fix left two columns. Set a conditional "left" css attribute, depend on if we select Chinese or not.

Risk:
We can't get the width of the first column, so the position of second column is hard coded. There's a bit movement when user scroll horizontally.

Reviewed by:
Kevin, Emily, Yujia, Annie, Dave