[CMLG-015] Refactor the code in Table and SearchPage components (#38)
Issue:

To allow users to choose between 10 rows/all and which page to look at, a Toggle Buttons component and a Table Pagination component need to pass information to the Table. After getting the data from the backend, Table also needs to pass how many pages are needed to the Pagination components. There are lots of props being passed around, also Table and Pagination are both passing props to each other, which seems a bit messy.

Solution:

Move all API calling methods to the SearchPage component, the resulted data is passed to the Table as a prop.
This leads to a new issue:

The Table originally contains data as a state and changes this state when sorting data. However, it cannot modify its props.

To solve that, the original sort function is split into two functions. Before displaying the data, a sort function is called to sort the data passed to the Table component based on the columnSortStatus state in Table. If user clicks a header, another function is called to change the columnSortStatus, which will trigger a re-rendering.

After change above, ToggleButton and Pagination components will both change the state of SearchPage based on user's action. New information is passed to all components by the SearchPage component.

Risk:
/

Reviewed by:
Annie, Yujia, Kevin, Dave, Eileen