[CMLG-015]  The toggle button of 10 rows and all rows in the search page.  (#46)
Issue:
Try to add a toggle button to the search page that allowed the user to choose what kind of format they want to display the translation table. One is 10 rows per page, the other one is showing all rows.

Solution: 
By using the button group that provided on the bootstrap library and using some css to make the toggle button have the same line as the other components like search Bar and selectCol button. And also make the colour consistence with the search page.

Risk: 
No risk currently.

Reviewed By: 
Emily, Dave, Kevin, Eileen and Annie.