[CMLG-013] Fix problems for select column/table and add hide hearder bar functionality to mobile view(#50)
Issue:
The user needs to click twice on the select column button to see the checklist. We can't see the horizontal scroll bar and use the mouse to scroll the table. And we need to click twice in the mobile view to hide the header bar, which is not a regular behavior, it should automatically close when we switch to other pages.

Solution: 
Use state to check if the select column component is clicked. Then use ref to store the current node, when the user clicks anywhere, the select column component will hide. Also use the ref for the mobile view header bar component, when the user clicks anywhere else the header bar will close. Change header bar scroll hidden to scroll-y hidden in order to fix the problem that the table can't scroll horizontally.

Risk:
Unknown

Reviewed by:
Kevin, Emily, Yujia, Annie, Dave