[CMLG-015]  Fixed the problem with the height of the table w/ pagination
Issue:
As the height of the table changes due to the result rows, the pagination moves with the the bottom of the table

Solution:
Changed search page to  be a flex container. gave the table a flex value.

Risk: Height of the search page is currently set to window height minus fixe header height.

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave