[CMLG-015]  Corrected the blue border selection bug
Issue:
Selected page will have the blue border on the wrong page when doing 5->7

Solution:
changed the keys for the items in the list to be reflective of the page number.
React will know to put the focus on the selected key that remains in the list(ie. the page that is selected)

Risk:
ELLIPSIS page item does not have a key assigned due to there being at most two instances, difficult to assign a unique key.
Not entirely sure if this will have any consquences.

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave