[CMLG-015] Table Pagination Styling (#65)
* [CMLG-015]  Added css for pagination

Issue:
The pagination currently has no styling.

Solution:
Add css for the pagination element and added css to the search page for the positioning of the pagination on the search page.

Risk:
none

Reviewed By: Emily, Dave, Kevin, Eileen and Annie.

* [CMLG-015]  Fixed the problem with the height of the table w/ pagination

Issue:
As the height of the table changes due to the result rows, the pagination moves with the the bottom of the table

Solution:
Changed search page to  be a flex container. gave the table a flex value.

Risk: Height of the search page is currently set to window height minus fixe header height.

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015] Fixed search page height responsiveness

Issue:
Search page height was fixed resulting in mobile mode having the pagination off the screen

Solution:
Changed root App div to  be a flex container.
Gave searchPage a flex value.
Made searchpage overflow-y: auto( somehow this fixes the responsive height problem with searchPage)

Risk: Pagination is not repsonsive as width cannot go lower than 340px.

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015]  Removed empty div spacing when showing all

Issue:
Small gap at the bottom of page when displaying all words and no pagination

Solution:
Nexted div with the if statement rather than the Pagination component

Risk: -

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015]  Made pagination responsive

Issue:
Pagination too wide for mobile devices creates awkward spacing on the right of the table

Solution:
Allow items to wrap to the next line

Risk: -

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015]  Dark mode for pagination

Issue:
New pagination has no dark mode styling

Solution:
Added css for darkmode for pagination
Removed some unnecessary css

Risk: -

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015]  Corrected the blue border selection bug

Issue:
Selected page will have the blue border on the wrong page when doing 5->7

Solution:
changed the keys for the items in the list to be reflective of the page number.
React will know to put the focus on the selected key that remains in the list(ie. the page that is selected)

Risk:
ELLIPSIS page item does not have a key assigned due to there being at most two instances, difficult to assign a unique key.
Not entirely sure if this will have any consquences.

Reviewed by: Emily, Yujia, Kevin, Eileen, Dave

* [CMLG-015] Implement hover for selected pages + small fixes

Issue:
Client would like to use the Helvetica font.

Solution:
Declare to use Helvetica globally.

Risk: No

Reviewed by: Emily, Yujia, Kevin, Annie, Dave