[CMLG-026]Solve the toggle button dark mode bug
Issue:
There is a bug with the active button. If you turn on on the dark mode after you click the "all" button, in dark mode, the "10 rows" button is highlighted instead of the "all" button. The details are stated in Github Issues #67

Solution:
Passing a state about "10 rows" and "all" from SearchPage to RowsPerPageToggleButton. 

Risk:
Highlight of button that toggles between "10 rows" and "all"

Reviewed by:
Annie, Eileen, Dave, Yujia, Emily